completed tasks on profile page, trailblazers
